feat(APP-381): Fix PR issues
- Separate component for DaoBreadcrumbs
- Separate utils for building breadcrumbs
