+++ .changeset/vast-moons-lose.md
@@ -2,4 +2,4 @@
 '@aragon/app': patch
 ---
 
-Improve the settings page according to the designs in Figma.
+Improve the settings page according to the designs in Figma
