+++ src/shared/components/daoBreadcrumbs/index.ts
@@ -1,0 +1,1 @@
+export { DaoBreadcrumbs, type IDaoBreadcrumbsProps } from './daoBreadcrumbs';
