+++ src/shared/components/processDataListItem/processDataListItem.tsx
@@ -1,10 +1,11 @@
 import type { IDao, IDaoPlugin } from '@/shared/api/daoService';
+import { DaoBreadcrumbs } from '@/shared/components/daoBreadcrumbs';
+import { daoBreadcrumbsUtils } from '@/shared/utils/daoBreadcrumbsUtils';
 import { daoUtils } from '@/shared/utils/daoUtils';
 import { subDaoDisplayUtils } from '@/shared/utils/subDaoDisplayUtils';
 import { DataList, type IDataListItemProps } from '@aragon/gov-ui-kit';
 import classNames from 'classnames';
 import { useTranslations } from '../translationsProvider';
-import { DaoBreadcrumbs, buildDaoBreadcrumbPath } from './daoBreadcrumbs';
 
 export type IProcessDataListItemProps = IDataListItemProps & {
     /**
@@ -28,7 +29,7 @@ export const ProcessDataListItem: React.FC<IProcessDataListItemProps> = (props) 
 
     const { address, description, slug } = process;
     const targetDaoAddress = subDaoDisplayUtils.getPluginDaoAddress(process);
-    const daoPath = buildDaoBreadcrumbPath({ rootDao: dao, targetAddress: targetDaoAddress });
+    const daoPath = daoBreadcrumbsUtils.buildDaoBreadcrumbPath({ rootDao: dao, targetAddress: targetDaoAddress });
 
     const processedDescription =
         description != null && description.length > 0
