+++ src/shared/utils/daoBreadcrumbsUtils/daoBreadcrumbsUtils.test.ts
@@ -1,17 +1,18 @@
 import type { ISubDaoSummary } from '@/shared/api/daoService';
 import { generateDao, generateSubDao } from '@/shared/testUtils';
-import { render, screen } from '@testing-library/react';
-import { buildDaoBreadcrumbPath, DaoBreadcrumbs } from './daoBreadcrumbs';
+import { daoBreadcrumbsUtils } from './daoBreadcrumbsUtils';
 
 describe('buildDaoBreadcrumbPath', () => {
     it('returns undefined when no dao or target is provided', () => {
-        expect(buildDaoBreadcrumbPath({ rootDao: undefined, targetAddress: undefined })).toBeUndefined();
+        expect(
+            daoBreadcrumbsUtils.buildDaoBreadcrumbPath({ rootDao: undefined, targetAddress: undefined }),
+        ).toBeUndefined();
     });
 
     it('returns the root DAO when target matches the root', () => {
         const dao = generateDao({ address: '0xparent', name: 'Parent DAO' });
 
-        const path = buildDaoBreadcrumbPath({ rootDao: dao, targetAddress: '0xparent' });
+        const path = daoBreadcrumbsUtils.buildDaoBreadcrumbPath({ rootDao: dao, targetAddress: '0xparent' });
 
         expect(path).toEqual([
             {
@@ -26,7 +27,7 @@ describe('buildDaoBreadcrumbPath', () => {
         const child = generateSubDao({ address: '0xchild', name: 'Child DAO' });
         const dao = generateDao({ address: '0xparent', name: 'Parent DAO', subDaos: [child] });
 
-        const path = buildDaoBreadcrumbPath({ rootDao: dao, targetAddress: '0xchild' });
+        const path = daoBreadcrumbsUtils.buildDaoBreadcrumbPath({ rootDao: dao, targetAddress: '0xchild' });
 
         expect(path?.map((node) => node.name)).toEqual(['Parent DAO', 'Child DAO']);
     });
@@ -39,31 +40,8 @@ describe('buildDaoBreadcrumbPath', () => {
         };
         const dao = generateDao({ address: '0xparent', name: 'Parent DAO', subDaos: [child] });
 
-        const path = buildDaoBreadcrumbPath({ rootDao: dao, targetAddress: '0xgrand' });
+        const path = daoBreadcrumbsUtils.buildDaoBreadcrumbPath({ rootDao: dao, targetAddress: '0xgrand' });
 
         expect(path?.map((node) => node.name)).toEqual(['Parent DAO', 'Child DAO', 'Grandchild DAO']);
-    });
-});
-
-describe('DaoBreadcrumbs', () => {
-    it('renders nothing when path is missing or too short', () => {
-        const { container } = render(<DaoBreadcrumbs path={undefined} />);
-        expect(container).toBeEmptyDOMElement();
-
-        const singlePath = [{ address: '0xparent', name: 'Parent', avatar: null }];
-        const { container: singleContainer } = render(<DaoBreadcrumbs path={singlePath} />);
-        expect(singleContainer).toBeEmptyDOMElement();
-    });
-
-    it('renders avatars and names for a breadcrumb chain', () => {
-        const path = [
-            { address: '0xparent', name: 'Parent DAO', avatar: null },
-            { address: '0xchild', name: 'Child DAO', avatar: null },
-        ];
-
-        render(<DaoBreadcrumbs path={path} />);
-
-        expect(screen.getByText('Parent DAO')).toBeInTheDocument();
-        expect(screen.getByText('Child DAO')).toBeInTheDocument();
     });
 });
