+++ src/shared/utils/daoBreadcrumbsUtils/index.ts
@@ -1,0 +1,1 @@
+export { daoBreadcrumbsUtils, type BreadcrumbNode, type IBuildDaoBreadcrumbPathParams } from './daoBreadcrumbsUtils';
